requirements.txt
pytest>=7.0.0
allure-pytest>=2.13.0
Pillow>=9.0.0
